View for Adding Comic to a Collection
Updated Main window to handle displaying the list of comics in a selected collection.
Added buttons to add and remove a comic from a collection and a context menu to remove a comic from a collection.
Created new FXML and codebehind for AddComic window.

Added TextFields for comic title and issue number, and buttons to confirm or cancel adding.