Binding View and ViewModel for Adding Comic to a Collection
Implemented MainWindow databinding and event handlers.
Bound comics list to the selected collection, added removal handler, and event to launch AddComicWindow with the ViewModel.
Implemented AddComicWindow databinding and event handlers.
Bound input fields and buttons for adding a comic and added event to close the window on add or cancel.